+++ src/extensions.js
@@ -30,12 +30,12 @@
 /*global modules, List, StageMorph, Costume, SpeechSynthesisUtterance, Sound,
 IDE_Morph, CamSnapshotDialogMorph, SoundRecorderDialogMorph, isSnapObject, nop,
 Color, Process, contains, localize, SnapTranslator, isString, detect, Point,
-SVG_Costume, newCanvas, WatcherMorph, BlockMorph, HatBlockMorph, invoke,
+SVG_Costume, newCanvas, WatcherMorph, BlockMorph, HatBlockMorph, invoke, isNil,
 BigUint64Array, DeviceOrientationEvent, DialogBoxMorph, Animation, console*/
 
 /*jshint esversion: 11, bitwise: false*/
 
-modules.extensions = '2025-November-20';
+modules.extensions = '2025-December-05';
 
 // Global stuff
 
@@ -1839,6 +1839,15 @@ SnapExtensions.primitives.set(
         ide = dlg ? dlg.ide : stage.parentThatIsA(IDE_Morph);
         if (!ide) {return ''; }
         return ide.currentSprite;
+    }
+);
+
+// Drawing on Sprites (pen_):
+
+SnapExtensions.primitives.set(
+    'pen_drawOn([sprite])',
+    function (sprite, proc) {
+        this.sheet = proc.reportObject(sprite);
     }
 );